+++ R/ClustImpute.R
@@ -84,13 +84,15 @@ ClustImpute <- function(X,nr_cluster, nr_iter=10, c_steps=1, wf=default_wf, n_en
                                       CENTROIDS=cl_old, seed_mode="keep_existing")
     }
 
-    # check if new centroids contain duplicates. If so, replace them by randomly drawn centroids
-    flag_dups = duplicated(cl_new)
-    nr_dups <-  sum(flag_dups)
+    # check if new centroids contain duplicate rows (min checks that indeed all values in a row are duplicates)
+    dupcliate_rows <- apply(matrix(duplicated(cl_new),nrow=dim(cl_new)[1],),1,min)
+    # If so, replace them by randomly drawn centroids
+    nr_dups <- sum(dupcliate_rows)
     while (nr_dups > 0)
     {
-      suppressWarnings(cl_new <- matrix(cl_new[!apply(cl_new,2,duplicated)],nr_cluster-nr_dups,dim(X)[2])) # keep unique centroids
-      # print(dim(cl_new))
+      # keep unique centroids
+      # suppressWarnings(cl_new <- matrix(cl_new[!apply(cl_new,2,duplicated)],nr_cluster-nr_dups,dim(X)[2]))
+      cl_new <- cl_new[!dupcliate_rows,]
       # get support for each column and draw uniformly from there
       sup_min <- apply(X,1,min)
       sup_max <- apply(X,1,max)
@@ -103,8 +105,8 @@ ClustImpute <- function(X,nr_cluster, nr_iter=10, c_steps=1, wf=default_wf, n_en
         new_centroids <- cbind(new_centroids, matrix(stats::runif(nr_dups,sup_min,sup_max),nr_dups,1))
       }
       cl_new <- rbind(cl_new,new_centroids) # add new to existing centroids
-      flag_dups = duplicated(cl_new)
-      nr_dups <-  sum(flag_dups)
+      dupcliate_rows <- apply(matrix(duplicated(cl_new),nrow=dim(cl_new)[1],),1,min)
+      nr_dups <- sum(dupcliate_rows)
     }
 
     # predict cluster
@@ -148,12 +150,15 @@ ClustImpute <- function(X,nr_cluster, nr_iter=10, c_steps=1, wf=default_wf, n_en
   cl_new <- ClusterR::KMeans_arma(data=X,clusters=nr_cluster,n_iter=c_steps,seed=seed_nr+n,
                                   CENTROIDS=cl_old, seed_mode="keep_existing")
 
-  # check if new centroids contain duplicates. If so, replace them by randomly drawn centroids
-  flag_dups = duplicated(cl_new)
-  nr_dups <-  sum(flag_dups)
+  # check if new centroids contain duplicate rows (min checks that indeed all values in a row are duplicates)
+  dupcliate_rows <- apply(matrix(duplicated(cl_new),nrow=dim(cl_new)[1],),1,min)
+  # If so, replace them by randomly drawn centroids
+  nr_dups <- sum(dupcliate_rows)
   while (nr_dups > 0)
   {
-    suppressWarnings(cl_new <- matrix(cl_new[!apply(cl_new,2,duplicated)],nr_cluster-nr_dups,dim(X)[2])) # keep unique centroids
+    # keep unique centroids
+    # suppressWarnings(cl_new <- matrix(cl_new[!apply(cl_new,2,duplicated)],nr_cluster-nr_dups,dim(X)[2]))
+    cl_new <- cl_new[!dupcliate_rows,]
     # get support for each column and draw uniformly from there
     sup_min <- apply(X,1,min)
     sup_max <- apply(X,1,max)
@@ -166,8 +171,8 @@ ClustImpute <- function(X,nr_cluster, nr_iter=10, c_steps=1, wf=default_wf, n_en
       new_centroids <- cbind(new_centroids, matrix(stats::runif(nr_dups,sup_min,sup_max),nr_dups,1))
     }
     cl_new <- rbind(cl_new,new_centroids) # add new to existing centroids
-    flag_dups = duplicated(cl_new)
-    nr_dups <-  sum(flag_dups)
+    dupcliate_rows <- apply(matrix(duplicated(cl_new),nrow=dim(cl_new)[1],),1,min)
+    nr_dups <- sum(dupcliate_rows)
   }
 
   # predict cluster one last time on final draws