+++ NEWS.md
@@ -1,7 +1,8 @@
 # ClustImpute 0.1.4
 
 * There is now a new option assign_with_wf. If set to False, then the weight function is only applied in the centroid computation, but ignored in the cluster assignment
-* Compatibility with new version of ClustImpute
+* New check for duplicate centroids. Duplicate rows are replaced with random draws.
+* Compatibility with new version 1.2.2 of ClusterR
 
 # ClustImpute 0.1.3